pyproject.toml
[tool.poetry]
name = "mlb_graphical_standings"
version = "1.0.0"
description = "Emails grpahical standings to me every Monday morning."
authors = ["Jordan Porter"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.2"
pybaseball = "^2.2.7"
tqdm = "^4.66.5"
chuy = "^1.4.0"
pylint = "^3.2.7"
mailtrap = "^2.0.1"
python-dotenv = "^1.0.1"
openai = "^1.43.0"

[tool.chuy]
build = "poetry build && && echo 'installing...' && poetry run pip install -q dist/mlb_graphical_standings-1.0.0-py3-none-any.whl --force-reinstall"
test = "poetry run pytest -s"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
